factory patern for creating wally nodes
