+++ wallynode/wallynode.go
@@ -3,7 +3,6 @@ package wallynode
 import (
 	"errors"
 	"fmt"
-	"github.com/hex0punk/wally/wallylib"
 	"golang.org/x/tools/go/callgraph"
 	"golang.org/x/tools/go/ssa"
 )
@@ -12,26 +11,37 @@ type WallyNode struct {
 	NodeString  string
 	Caller      *callgraph.Node
 	Site        ssa.CallInstruction
-	Recoverable bool
+	recoverable bool
 }
 
-func NewWallyNode(nodeStr string, caller *callgraph.Node, site ssa.CallInstruction, connectedNodes map[*ssa.Function]*callgraph.Node) WallyNode {
-	recoverable := false
-	if nodeStr == "" {
-		if site == nil {
-			nodeStr = fmt.Sprintf("Func: %s.[%s] %s", caller.Func.Pkg.Pkg.Name(), caller.Func.Name(), wallylib.GetFormattedPos(caller.Func.Package(), caller.Func.Pos()))
-		} else {
-			fp := wallylib.GetFormattedPos(caller.Func.Package(), site.Pos())
-			recoverable = IsRecoverable(caller, connectedNodes)
-			nodeStr = GetNodeString(fp, caller, recoverable)
-		}
-	}
-	return WallyNode{
-		NodeString:  nodeStr,
-		Caller:      caller,
-		Site:        site,
-		Recoverable: recoverable,
-	}
+type NodeType int
+
+const (
+	Site NodeType = iota
+	Function
+)
+
+//func NewWallyNode(nodeStr string, caller *callgraph.Node, site ssa.CallInstruction, connectedNodes map[*ssa.Function]*callgraph.Node) WallyNode {
+//	recoverable := false
+//	if nodeStr == "" {
+//		if site == nil {
+//			nodeStr = fmt.Sprintf("Func: %s.[%s] %s", caller.Func.Pkg.Pkg.Name(), caller.Func.Name(), wallylib.GetFormattedPos(caller.Func.Package(), caller.Func.Pos()))
+//		} else {
+//			fp := wallylib.GetFormattedPos(caller.Func.Package(), site.Pos())
+//			recoverable = IsRecoverable(caller, connectedNodes)
+//			nodeStr = GetNodeString(fp, caller, recoverable)
+//		}
+//	}
+//	return WallyNode{
+//		NodeString:  nodeStr,
+//		Caller:      caller,
+//		Site:        site,
+//		recoverable: recoverable,
+//	}
+//}
+
+func (n *WallyNode) IsRecoverable() bool {
+	return n.recoverable
 }
 
 func GetNodeString(basePos string, s *callgraph.Node, recoverable bool) string {
@@ -44,36 +54,6 @@ func GetNodeString(basePos string, s *callgraph.Node, recoverable bool) string {
 	}
 
 	return baseStr
-}
-
-func IsRecoverable(s *callgraph.Node, callgraphNodes map[*ssa.Function]*callgraph.Node) bool {
-	function := s.Func
-	if function.Recover != nil {
-		rec, err := findDeferRecover(function, function.Recover.Index-1)
-		if err == nil && rec {
-			return true
-		}
-	}
-	if wallylib.IsClosure(function) {
-		enclosingFunc := closureArgumentOf(s, callgraphNodes[s.Func.Parent()])
-		if enclosingFunc != nil && enclosingFunc.Recover != nil {
-			rec, err := findDeferRecover(enclosingFunc, enclosingFunc.Recover.Index-1)
-			if err == nil && rec {
-				return true
-			}
-		}
-		if enclosingFunc != nil {
-			for _, af := range enclosingFunc.AnonFuncs {
-				if af.Recover != nil {
-					rec, err := findDeferRecover(af, af.Recover.Index-1)
-					if err == nil && rec {
-						return true
-					}
-				}
-			}
-		}
-	}
-	return false
 }
 
 func findDeferRecover(fn *ssa.Function, idx int) (bool, error) {
